refactor driver
