+++ onnx-plugin/CMakeLists.txt
@@ -25,6 +25,7 @@ message(STATUS "ONNX Runtime lib dir: ${ONNXRUNTIME_LIB_DIR}")
 
 add_library(tinyxpu_ep SHARED
     src/tinyxpu_ep.cpp
+    src/tinyxpu_driver.cpp
     src/tinyxpu_perf.cpp
 )
